v1.104:
- NRI: eliminated FrameBuffer (VK uses dynamic rendering)
- NRI: introduced "Sample_t" type to mimic "uint8_t" in implementations
- NRI: tweaked types in descs
- VK: fixed validation error in Fence
- introduced .clang-format (currently applied only to Validation)
- updated doc

+++ Source/D3D11/PipelineD3D11.h
@@ -82,8 +82,8 @@ private:
 
     D3D11_PRIMITIVE_TOPOLOGY m_Topology = D3D11_PRIMITIVE_TOPOLOGY_UNDEFINED;
     Color32f m_BlendFactor = {};
-    uint16_t m_SampleMask = 0;
-    uint16_t m_StencilRef = uint16_t(-1);
+    uint32_t m_SampleMask = 0;
+    uint8_t m_StencilRef = uint8_t(-1);
     bool m_IsRasterizerDiscarded = false;
     const PipelineLayoutD3D11* m_PipelineLayout = nullptr;
 };
